working with builder, application complete

+++ src/com/company/Addition.java
@@ -3,7 +3,7 @@ package com.company;
 public class Addition implements Strategy{
 
     @Override
-    public float calculation(float a, float b) {
+    public int calculation(int a, int b) {
         return a+b;
     }
